+++ src/main/java/com/mns/cda/suivimns/service/entity/CommentService.java
@@ -4,10 +4,14 @@ import com.mns.cda.suivimns.dao.CommentDao;
 import com.mns.cda.suivimns.dto.entity.CommentDto;
 import com.mns.cda.suivimns.mapper.entity.CommentMapper;
 import com.mns.cda.suivimns.model.Comment;
+import com.mns.cda.suivimns.security.AppUserDetails;
 import lombok.RequiredArgsConstructor;
+import org.springframework.http.HttpStatus;
+import org.springframework.http.ResponseEntity;
 import org.springframework.stereotype.Service;
 
 import java.util.List;
+import java.util.Objects;
 
 @Service
 @RequiredArgsConstructor
@@ -16,6 +20,8 @@ public class CommentService  {
 
     public static class CommentNotFoundException extends RuntimeException {
     }
+
+    public static class CommentNotOwnedException extends RuntimeException {}
 
     protected final CommentDao commentDao;
     protected final CommentMapper commentMapper;
@@ -39,17 +45,29 @@ public class CommentService  {
         return commentMapper.toDto(saved);
     }
 
-    public void delete(int id) throws CommentService.CommentNotFoundException {
+    public void delete(int id, AppUserDetails userDetails) throws CommentService.CommentNotFoundException {
         Comment comment = commentDao.findById(id)
                 .orElseThrow(CommentService.CommentNotFoundException::new);
+
+        // On verifie si l'utilisateur est admin ou s'il est le proprietaire de la ressource
+        if (!Objects.equals(userDetails.getUserRole(), "ADMIN") &&
+                !comment.getAuthor().getIdAppUser().equals(userDetails.getId())) {
+            throw new CommentNotOwnedException();
+        }
 
         commentDao.delete(comment);
     }
 
-    public CommentDto update(int id, CommentDto commentToUpdate) throws CommentService.CommentNotFoundException {
+    public CommentDto update(int id, CommentDto commentToUpdate, AppUserDetails userDetails) throws CommentService.CommentNotFoundException {
 
         Comment currentComment = commentDao.findById(id)
                 .orElseThrow(CommentService.CommentNotFoundException::new);
+
+        // On verifie si l'utilisateur est admin ou s'il est le proprietaire de la ressource
+        if (!Objects.equals(userDetails.getUserRole(), "ADMIN") &&
+                !currentComment.getAuthor().getIdAppUser().equals(userDetails.getId())) {
+            throw new CommentNotOwnedException();
+        }
 
         commentMapper.updateEntityFromDto(commentToUpdate, currentComment);
