+++ src/main/java/com/mns/cda/suivimns/service/entity/ManagerService.java
@@ -7,6 +7,7 @@ import com.mns.cda.suivimns.dto.flat.PasswordDto;
 import com.mns.cda.suivimns.mapper.entity.ManagerMapper;
 import com.mns.cda.suivimns.model.Client;
 import com.mns.cda.suivimns.model.Manager;
+import com.mns.cda.suivimns.security.AppUserDetails;
 import lombok.RequiredArgsConstructor;
 import org.springframework.security.crypto.password.PasswordEncoder;
 import org.springframework.stereotype.Service;
@@ -57,18 +58,28 @@ public class ManagerService  {
         appUserDao.save(manager);
     }
 
-    public void delete(int id) throws ManagerService.ManagerNotFoundException {
+    public void delete(int id, AppUserDetails userDetails) throws ManagerService.ManagerNotFoundException, AppUserService.AccountNotOwnedException {
         Manager manager = managerDao.findById(id)
                 .orElseThrow(ManagerService.ManagerNotFoundException::new);
+
+        if (!Objects.equals(userDetails.getUserRole(), "ADMIN") &&
+                userDetails.getId() != id) {
+            throw new AppUserService.AccountNotOwnedException();
+        }
 
         managerDao.delete(manager);
     }
 
-    public ManagerDto update(int id, ManagerDto dto)
-            throws ManagerService.ManagerNotFoundException, AppUserService.EmailAlreadyUsedException {
+    public ManagerDto update(int id, ManagerDto dto, AppUserDetails userDetails)
+            throws ManagerService.ManagerNotFoundException, AppUserService.EmailAlreadyUsedException, AppUserService.AccountNotOwnedException {
 
         if (appUserDao.existsByEmail(dto.email())) {
             throw new AppUserService.EmailAlreadyUsedException();
+        }
+
+        if (!Objects.equals(userDetails.getUserRole(), "ADMIN") &&
+                userDetails.getId() != id) {
+            throw new AppUserService.AccountNotOwnedException();
         }
 
         Manager currentManager = managerDao.findById(id)
@@ -79,8 +90,8 @@ public class ManagerService  {
         return managerMapper.toDto(managerDao.save(currentManager));
     }
 
-    public void updatePassword(int id, PasswordDto dto)
-            throws ManagerService.ManagerNotFoundException, ManagerService.BadPasswordException {
+    public void updatePassword(int id, PasswordDto dto, AppUserDetails userDetails)
+            throws ManagerService.ManagerNotFoundException, ManagerService.BadPasswordException, AppUserService.AccountNotOwnedException {
 
         Manager user = managerDao.findById(id)
                 .orElseThrow(ManagerService.ManagerNotFoundException::new);
@@ -88,6 +99,11 @@ public class ManagerService  {
         // vérifier ancien mot de passe
         if (!Objects.equals(user.getPassword(), dto.oldPassword())) {
             throw new ManagerService.BadPasswordException();
+        }
+
+        if (!Objects.equals(userDetails.getUserRole(), "ADMIN") &&
+                userDetails.getId() != id) {
+            throw new AppUserService.AccountNotOwnedException();
         }
 
         user.setPassword(dto.newPassword());
